grafana-as-code/requirements.txt
# Grafana API Client dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Optional: For advanced features
prometheus-client>=0.19.0
python-dateutil>=2.8.2

# Development
pytest>=7.4.3
pytest-mock>=3.12.0
black>=23.12.0
